+++ Rdmp.Core/CommandExecution/AtomicCommands/ExecuteCommandDeprecate.cs
@@ -5,23 +5,21 @@
 // You should have received a copy of the GNU General Public License along with RDMP. If not, see <https://www.gnu.org/licenses/>.
 
 using MapsDirectlyToDatabaseTable;
-using Rdmp.Core.CommandExecution.AtomicCommands;
 using Rdmp.Core.Curation.Data;
-using Rdmp.UI.ItemActivation;
 
-namespace Rdmp.UI.CommandExecution.AtomicCommands
+namespace Rdmp.Core.CommandExecution.AtomicCommands
 {
-    public class ExecuteCommandDeprecate : BasicUICommandExecution,IAtomicCommand
+    public class ExecuteCommandDeprecate : BasicCommandExecution, IAtomicCommand
     {
         private readonly IMightBeDeprecated _o;
         private bool _desiredState;
 
-        public ExecuteCommandDeprecate(IActivateItems itemActivator, IMightBeDeprecated o) : this(itemActivator,o,true)
+        public ExecuteCommandDeprecate(IBasicActivateItems itemActivator, IMightBeDeprecated o) : this(itemActivator, o, true)
         {
-            
+
         }
 
-        public ExecuteCommandDeprecate(IActivateItems itemActivator, IMightBeDeprecated o, bool desiredState) : base(itemActivator)
+        public ExecuteCommandDeprecate(IBasicActivateItems itemActivator, IMightBeDeprecated o, bool desiredState) : base(itemActivator)
         {
             _o = o;
             _desiredState = desiredState;